fitsnap-ai/backend/requirements.txt
# FastAPI & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
hiredis==2.2.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# AI & Vision
openai==1.6.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2

# Fashion & Retail APIs
requests==2.31.0
aiohttp==3.9.1

# Image Processing
scikit-image==0.22.0
colorthief==0.2.1

# Payments
stripe==7.8.0

# Email
python-multipart==0.0.6
aiosmtplib==3.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3